+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Melissa HTTP Server Changelog
 
+### v0.1
+- Initial version with basic implementations of
+	- File operations
+	- Server response headers
+	- Request headers parsing
+
+
 ### v0.0.1
 - Very early prototype with early sockets code and dummy response.